Fixed authorization issues for Api Controllers

+++ Kore.Web.ContentManagement/Areas/Admin/Localization/Controllers/Api/LanguageApiController.cs
@@ -1,18 +1,21 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Net.Http.Headers;
 using System.Web.Http;
 using System.Web.Http.OData;
+using System.Web.Http.Results;
 using Kore.Caching;
 using Kore.Data;
 using Kore.Infrastructure;
 using Kore.Localization;
 using Kore.Localization.Domain;
 using Kore.Web.Http.OData;
+using Kore.Web.Security.Membership.Permissions;
 using LanguageEntity = Kore.Localization.Domain.Language;
 
 namespace Kore.Web.ContentManagement.Areas.Admin.Localization.Controllers.Api
 {
-    [Authorize(Roles = KoreConstants.Roles.Administrators)]
+    //[Authorize(Roles = KoreConstants.Roles.Administrators)]
     public class LanguageApiController : GenericODataController<LanguageEntity, Guid>
     {
         private readonly Lazy<ICacheManager> cacheManager;
@@ -41,6 +44,11 @@ namespace Kore.Web.ContentManagement.Areas.Admin.Localization.Controllers.Api
         [HttpPost]
         public virtual IHttpActionResult ResetLocalizableStrings(ODataActionParameters parameters)
         {
+            if (!CheckPermission(WritePermission))
+            {
+                return new UnauthorizedResult(new AuthenticationHeaderValue[0], ActionContext.Request);
+            }
+
             localizableStringRepository.Value.DeleteAll();
 
             var localizedStringsProviders = EngineContext.Current.ResolveAll<IDefaultLocalizableStringsProvider>();
@@ -67,6 +75,16 @@ namespace Kore.Web.ContentManagement.Areas.Admin.Localization.Controllers.Api
             cacheManager.Value.Remove("LocalizableStrings_");
 
             return Ok();
+        }
+
+        protected override Permission ReadPermission
+        {
+            get { return CmsPermissions.LanguagesRead; }
+        }
+
+        protected override Permission WritePermission
+        {
+            get { return CmsPermissions.LanguagesWrite; }
         }
     }
 }
